Revert "feature [#19] di 사용 전 세팅, module설정"
This reverts commit d74bb65f3d847ca2966f87f46cbe5a5caf9e9ac7.
