--- app/src/main/java/com/example/commerceapp/MainApplication.kt
@@ -1,8 +1,0 @@
-package com.example.commerceapp
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class MainApplication : Application() {
-}
